refactor: context 폴더 분리
